LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/the-select-element/customizable-select/selectedcontent-mutations-expected.txt










PASS MutationObserver records during parsing of <select> with <selectedcontent>: test1
PASS MutationObserver records during parsing of <select> with <selectedcontent>: test2
FAIL MutationObserver records during parsing of <select> with <selectedcontent>: test3 assert_array_equals: test3 lengths differ, expected array […, "Type: childList | Target: selectedcontent | Added: [#text: \"inner1\"]", "Type: childList | Target: selectedcontent | Added: [#text: \"\"] | After: selectedcontent", "Type: childList | Target: button | Added: [#text: \"\"] | After: selectedcontent", "Type: childList | Target: button | Added: [selectedcontent] | After: #text: \"\"", "Type: childList | Target: selectedcontent | Added: [#text: \"outer2\"]", "Type: childList | Target: selectedcontent | Added: [selectedcontent] | After: #text: \"outer2\"", "Type: childList | Target: selectedcontent | Added: [#text: \"inner2\"]", "Type: childList | Target: selectedcontent | Added: [#text: \"\"] | After: selectedcontent", "Type: childList | Target: button | Added: [#text: \"\"] | After: selectedcontent", "Type: childList | Target: select | Added: [#text: \"\"] | After: button", "Type: childList | Target: select | Added: [option] | After: #text: \"\"", "Type: childList | Target: selectedcontent | Removed: [#text: \"outer1\", selectedcontent, #text: \"\"]", "Type: childList | Target: selectedcontent | Removed: [#text: \"inner1\"]", "Type: childList | Target: selectedcontent | Removed: [#text: \"outer2\", selectedcontent, #text: \"\"]", "Type: childList | Target: selectedcontent | Removed: [#text: \"inner2\"]", "Type: childList | Target: option | Added: [#text: \"option\"]", "Type: childList | Target: selectedcontent | Added: [#text: \"option\"]", "Type: childList | Target: selectedcontent | Added: [#text: \"option\"]", "Type: childList | Target: select | Added: [#text: \"\"] | After: option", "Type: childList | Target: div#test3 | Added: [#text: \"\"] | After: select"] length 28, got […, "Type: childList | Target: button | Added: [selectedcontent] | After: #text: \"\"", "Type: childList | Target: selectedcontent | Added: [#text: \"outer1\"]", "Type: childList | Target: selectedcontent | Added: [selectedcontent] | After: #text: \"outer1\"", "Type: childList | Target: selectedcontent | Added: [#text: \"inner1\"]", "Type: childList | Target: selectedcontent | Added: [#text: \"\"] | After: selectedcontent", "Type: childList | Target: button | Added: [#text: \"\"] | After: selectedcontent", "Type: childList | Target: button | Added: [selectedcontent] | After: #text: \"\"", "Type: childList | Target: selectedcontent | Removed: [#text: \"outer1\", selectedcontent, #text: \"\"]", "Type: childList | Target: selectedcontent | Added: [#text: \"outer2\"]", "Type: childList | Target: selectedcontent | Added: [selectedcontent] | After: #text: \"outer2\"", "Type: childList | Target: selectedcontent | Added: [#text: \"inner2\"]", "Type: childList | Target: selectedcontent | Added: [#text: \"\"] | After: selectedcontent", "Type: childList | Target: button | Added: [#text: \"\"] | After: selectedcontent", "Type: childList | Target: select | Added: [#text: \"\"] | After: button", "Type: childList | Target: select | Added: [option] | After: #text: \"\"", "Type: childList | Target: option | Added: [#text: \"option\"]", "Type: childList | Target: selectedcontent | Added: [#text: \"option\"]", "Type: childList | Target: selectedcontent | Added: [#text: \"option\"] | Removed: [#text: \"outer2\", selectedcontent, #text: \"\"]", "Type: childList | Target: select | Added: [#text: \"\"] | After: option", "Type: childList | Target: div#test3 | Added: [#text: \"\"] | After: select"] length 25
FAIL MutationObserver records during parsing of <select> with <selectedcontent>: test4 assert_array_equals: test4 lengths differ, expected array […, "Type: childList | Target: button | Added: [#text: \"\"]", "Type: childList | Target: button | Added: [selectedcontent] | After: #text: \"\"", "Type: childList | Target: button | Added: [#text: \"\"] | After: selectedcontent", "Type: childList | Target: select | Added: [#text: \"\"] | After: button", "Type: childList | Target: select | Added: [div] | After: #text: \"\"", "Type: childList | Target: div | Added: [#text: \"\"]", "Type: childList | Target: div | Added: [option] | After: #text: \"\"", "Type: childList | Target: option | Added: [#text: \"one\"]", "Type: childList | Target: selectedcontent | Added: [#text: \"one\"]", "Type: childList | Target: div | Added: [#text: \"\"] | After: option", "Type: childList | Target: select | Added: [#text: \"\"] | After: div", "Type: childList | Target: select | Added: [div] | After: #text: \"\"", "Type: childList | Target: div | Added: [#text: \"\"]", "Type: childList | Target: div | Added: [option] | After: #text: \"\"", "Type: childList | Target: selectedcontent | Removed: [#text: \"one\"]", "Type: childList | Target: option | Added: [#text: \"two\"]", "Type: childList | Target: selectedcontent | Added: [#text: \"two\"]", "Type: childList | Target: div | Added: [#text: \"\"] | After: option", "Type: childList | Target: select | Added: [#text: \"\"] | After: div", "Type: childList | Target: div#test4 | Added: [#text: \"\"] | After: select"] length 24, got […, "Type: childList | Target: select | Added: [button] | After: #text: \"\"", "Type: childList | Target: button | Added: [#text: \"\"]", "Type: childList | Target: button | Added: [selectedcontent] | After: #text: \"\"", "Type: childList | Target: button | Added: [#text: \"\"] | After: selectedcontent", "Type: childList | Target: select | Added: [#text: \"\"] | After: button", "Type: childList | Target: select | Added: [div] | After: #text: \"\"", "Type: childList | Target: div | Added: [#text: \"\"]", "Type: childList | Target: div | Added: [option] | After: #text: \"\"", "Type: childList | Target: option | Added: [#text: \"one\"]", "Type: childList | Target: selectedcontent | Added: [#text: \"one\"]", "Type: childList | Target: div | Added: [#text: \"\"] | After: option", "Type: childList | Target: select | Added: [#text: \"\"] | After: div", "Type: childList | Target: select | Added: [div] | After: #text: \"\"", "Type: childList | Target: div | Added: [#text: \"\"]", "Type: childList | Target: div | Added: [option] | After: #text: \"\"", "Type: childList | Target: option | Added: [#text: \"two\"]", "Type: childList | Target: selectedcontent | Added: [#text: \"two\"] | Removed: [#text: \"one\"]", "Type: childList | Target: div | Added: [#text: \"\"] | After: option", "Type: childList | Target: select | Added: [#text: \"\"] | After: div", "Type: childList | Target: div#test4 | Added: [#text: \"\"] | After: select"] length 23
PASS MutationObserver records during parsing of <select> with <selectedcontent>: test5
PASS MutationObserver records during parsing of <select> with <selectedcontent>: test6
PASS MutationObserver records during parsing of <select> with <selectedcontent>: test7
FAIL MutationObserver records during parsing of <select> with <selectedcontent>: test8 assert_array_equals: test8 expected property 19 to be "Type: childList | Target: div.container | Added: [select]" but got "Type: childList | Target: selectedcontent | Added: [#text: \"one\"] | Removed: [#text: \"one\"]" (expected array ["Type: childList | Target: div#test8 | Added: [#text: \"\"] | After: script", "Type: childList | Target: div#test8 | Added: [div.container] | After: #text: \"\"", "Type: childList | Target: div#test8 | Added: [#text: \"\"] | After: div.container", "Type: childList | Target: div#test8 | Added: [select] | After: #text: \"\"", "Type: childList | Target: select | Added: [#text: \"\"]", "Type: childList | Target: select | Added: [button] | After: #text: \"\"", "Type: childList | Target: button | Added: [#text: \"\"]", "Type: childList | Target: button | Added: [selectedcontent] | After: #text: \"\"", "Type: childList | Target: button | Added: [#text: \"\"] | After: selectedcontent", "Type: childList | Target: select | Added: [#text: \"\"] | After: button", "Type: childList | Target: select | Added: [option] | After: #text: \"\"", "Type: childList | Target: option | Added: [#text: \"one\"]", "Type: childList | Target: selectedcontent | Added: [#text: \"one\"]", "Type: childList | Target: select | Added: [#text: \"\"] | After: option", "Type: childList | Target: select | Added: [option] | After: #text: \"\"", "Type: childList | Target: option | Added: [#text: \"two\"]", "Type: childList | Target: select | Added: [#text: \"\"] | After: option", "Type: childList | Target: div#test8 | Added: [#text: \"\"] | After: select", "Type: childList | Target: div#test8 | Removed: [select] | After: #text: \"\" | Before: #text: \"\"", "Type: childList | Target: div.container | Added: [select]", "Type: childList | Target: selectedcontent | Added: [#text: \"one\"] | Removed: [#text: \"one\"]"] got ["Type: childList | Target: div#test8 | Added: [#text: \"\"] | After: script", "Type: childList | Target: div#test8 | Added: [div.container] | After: #text: \"\"", "Type: childList | Target: div#test8 | Added: [#text: \"\"] | After: div.container", "Type: childList | Target: div#test8 | Added: [select] | After: #text: \"\"", "Type: childList | Target: select | Added: [#text: \"\"]", "Type: childList | Target: select | Added: [button] | After: #text: \"\"", "Type: childList | Target: button | Added: [#text: \"\"]", "Type: childList | Target: button | Added: [selectedcontent] | After: #text: \"\"", "Type: childList | Target: button | Added: [#text: \"\"] | After: selectedcontent", "Type: childList | Target: select | Added: [#text: \"\"] | After: button", "Type: childList | Target: select | Added: [option] | After: #text: \"\"", "Type: childList | Target: option | Added: [#text: \"one\"]", "Type: childList | Target: selectedcontent | Added: [#text: \"one\"]", "Type: childList | Target: select | Added: [#text: \"\"] | After: option", "Type: childList | Target: select | Added: [option] | After: #text: \"\"", "Type: childList | Target: option | Added: [#text: \"two\"]", "Type: childList | Target: select | Added: [#text: \"\"] | After: option", "Type: childList | Target: div#test8 | Added: [#text: \"\"] | After: select", "Type: childList | Target: div#test8 | Removed: [select] | After: #text: \"\" | Before: #text: \"\"", "Type: childList | Target: selectedcontent | Added: [#text: \"one\"] | Removed: [#text: \"one\"]", "Type: childList | Target: div.container | Added: [select]"])
